+++ packages/core/botpress-builtins/src/setup.js
@@ -18,13 +18,20 @@ export default function(bp) {
   bp.dialogEngine.onBeforeCreated(async (ctx, next) => {
     const { stateId } = ctx
 
-    if (!stateId.includes(':')) {
+    if (!stateId || !stateId.includes(':')) {
       // Unknown platform / can't extract userId
       return next()
     }
 
-    const convoCount = await bp.users.getTag(stateId, USER_TAG_CONVO_COUNT)
-    await bp.users.tag(stateId, USER_TAG_CONVO_COUNT, parseInt(convoCount || 0) + 1)
+    // TODO: implement proper atomic increment
+    // and eliminate simultaneous INSERTs that violate
+    // unique constraint
+    try {
+      const convoCount = await bp.users.getTag(stateId, USER_TAG_CONVO_COUNT)
+      await bp.users.tag(stateId, USER_TAG_CONVO_COUNT, parseInt(convoCount || 0) + 1)
+    } catch (err) {
+      console.error(err.message)
+    }
 
     next()
   })
@@ -36,13 +43,19 @@ export default function(bp) {
   bp.dialogEngine.onBeforeEnd(async (ctx, next) => {
     const { stateId } = ctx
 
-    if (!stateId.includes(':')) {
+    if (!stateId || !stateId.includes(':')) {
       // Unknown platform / can't extract userId
       return next()
     }
 
-    const position = await bp.dialogEngine.getCurrentPosition(stateId)
-    await bp.users.tag(stateId, USER_TAG_CONVO_LAST, position && position.flow)
+    // TODO: eliminate simultaneous INSERTs that violate
+    // unique constraint
+    try {
+      const position = await bp.dialogEngine.getCurrentPosition(stateId)
+      await bp.users.tag(stateId, USER_TAG_CONVO_LAST, position && position.flow)
+    } catch (err) {
+      console.error(err.message)
+    }
 
     //
     // Cleans up Conversation Storage variables